requirements.txt
alembic==1.12.1
annotated-types==0.6.0
anyio==3.7.1
astroid==3.0.1
bcrypt==4.1.1
black==23.11.0
certifi==2023.7.22
cffi==1.16.0
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==41.0.7
dill==0.3.7
distlib==0.3.7
dnspython==2.4.2
ecdsa==0.18.0
email-validator==2.1.0.post1
exceptiongroup==1.1.3
fastapi==0.104.1
filelock==3.13.1
ghp-import==2.1.0
greenlet==3.0.1
griffe==0.38.0
h11==0.14.0
httptools==0.6.1
identify==2.5.31
idna==3.4
iniconfig==2.0.0
isort==5.12.0
Jinja2==3.1.2
Mako==1.3.0
Markdown==3.5.1
MarkupSafe==2.1.3
mccabe==0.7.0
mergedeep==1.3.4
mkdocs==1.5.3
mkdocs-autorefs==0.5.0
mkdocstrings==0.24.0
mkdocstrings-python==1.7.4
mypy-extensions==1.0.0
nodeenv==1.8.0
packaging==23.2
passlib==1.7.4
pathspec==0.11.2
platformdirs==3.11.0
pluggy==1.3.0
pre-commit==3.5.0
pyasn1==0.5.1
pycparser==2.21
pydantic==2.5.1
pydantic_core==2.14.3
pylint==3.0.2
pymdown-extensions==10.4
pytest==7.4.3
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
python-jose==3.3.0
PyYAML==6.0.1
pyyaml_env_tag==0.1
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.23
starlette==0.27.0
tomli==2.0.1
tomlkit==0.12.2
typing_extensions==4.8.0
urllib3==2.0.7
uvicorn==0.24.0.post1
virtualenv==20.24.6
watchdog==3.0.0
watchfiles==0.21.0
websockets==12.0
